Add files via upload
This code is written for TM4C123GH6PM using Thumb instructions in assembly programming language via Keil uVision5.
OutStr.s and Startup.s are printing to the terminal via UART and initialization of the MCU respectively, they are essentials which are not written by me.
In main.s, I simply load a number to R4 register and I hold an address in R5 register. Then I call my subroutine CNVRT.s and call OutStr.s subroutine to print it.
CNVRT.s is taking the value in R4 and find its decimal equivalent, and print it in a proper format. 
For example, R4= 0x00A2 --> output: 162 or R4= 0X0001 --> output: 1 and so on.